+++ packages/gov-ui-kit/src/modules/components/address/addressInput/index.ts
@@ -1,1 +1,1 @@
-export { AddressInput, type IAddressInputProps } from './addressInput';
+export { AddressInput, type IAddressInputProps, type IAddressInputResolvedValue } from './addressInput';