+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   Bump `softprops/action-gh-release` from 2.0.8 to 2.0.9
+
 ## [1.0.52] - 2024-10-30
 
 ### Changed